Remove Monoid instance
There is no good way to define a Monoid instance for Validation that
retains failures. The two identity rules required for Monoid instances
are:

- mempty <> x = x
- x <> mempty = x

There is only one possible definitions of mempty that meets these
requirements, and preserves errors:

- mempty = Success mempty

However, this would require that the Right side of Validation have a
Monoid instance. Someone could add this instance with a Monoid a
constaint, however, it is not so useful because the Right side is
usually not a Monoid.