Simplify prism/iso stuff
There was a good bit of cruft I'm pretty sure wasn't needed.
However, inlining and such may change, and this should be
investigated and dealt with. Also, for "expensive" functors,
it may pay to play tricks to fuse `fmap`s. I *think* this can be
done by using `clonePrism (a . b)` rather than just `a . b` in
various places.